WIP: The test suite has address sanitiser issues.
For some its that the DOM test code expects nodes returned by libdom
APIs to always be reffed.  So this change refs them.

For test_level1_html_HTMLSelectElement08, the problem is that in the
while loop, it doesn't unref `voption` at the end of the loop, in the
test code.

For test_level1_html_HTMLTableElement07, the problem is the
while loop doesn't unref `vrow` at the end of the loop in the
test code.

For test_level1_html_HTMLTableElement09, it's the same for
variable, `vbodies`.